requirements.txt
aniso8601==1.3.0
asn1crypto==0.23.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
boto==2.48.0
bz2file==0.98
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
Fabric3==1.13.1.post1
Flask==0.12.2
Flask-RESTful==0.3.6
gensim==3.1.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
nltk==3.2.5
numpy==1.13.3
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.0
PyStemmer==1.3.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
smart-open==1.5.3
textblob==0.14.0
urllib3==1.22
Werkzeug==0.12.2
